hash

+++ script_delete.php
@@ -1,26 +1,27 @@
 <?php
-ini_set('error_reporting', E_ALL);
-ini_set('display_errors', 1);
-ini_set('display_startup_errors', 1);
-include("dbconnect.php");
-$c_mass = $_POST['checkbox_mass'];
-$name_t = $_POST['selected_table']; // Имя нужной таблицы
-if(empty($c_mass))
+session_start();
+if ($_SESSION["is_auth"] == true && $_SESSION["type"] == "admin")
 {
-    header("Location: ".$_SERVER['HTTP_REFERER']);
-    exit();
-}
-else
-{
-    if ($name_t == "posts" or $name_t == "users")
+    include("dbconnect.php");
+    $c_mass = $_POST['checkbox_mass'];
+    $name_t = $_POST['selected_table']; // Имя нужной таблицы
+    if(empty($c_mass))
     {
-        foreach ($c_mass as $row)
-        {
-            global $pdo;
-            $query = $pdo->prepare("DELETE FROM $name_t WHERE id=:id");
-            $query->execute(array("id" => $row));
-        }
+        header("Location: ".$_SERVER['HTTP_REFERER']);
+        exit();
     }
-    header("Location: ".$_SERVER['HTTP_REFERER']);
-    exit();
+    else
+    {
+        if ($name_t == "posts" or $name_t == "users")
+        {
+            foreach ($c_mass as $row)
+            {
+                global $pdo;
+                $query = $pdo->prepare("DELETE FROM $name_t WHERE id=:id");
+                $query->execute(array("id" => $row));
+            }
+        }
+        header("Location: ".$_SERVER['HTTP_REFERER']);
+        exit();
+    }
 }
